Separating out form and image preview from Image component
