+++ frontend/src/components/NavBar/PrintImage/preview.tsx
@@ -1,0 +1,374 @@
+import {
+  Box,
+  Typography,
+  WithStyles,
+  createStyles,
+  withStyles,
+} from '@material-ui/core';
+import maplibregl from 'maplibre-gl';
+import React, { useRef } from 'react';
+import MapGL, { Layer, MapRef, Marker, Source } from 'react-map-gl/maplibre';
+import { useSelector } from 'react-redux';
+import { mapStyle } from 'components/MapView/Map';
+import { getFormattedDate } from 'utils/date-utils';
+import { appConfig } from 'config';
+import { AdminCodeString } from 'config/types';
+import { lightGrey } from 'muiTheme';
+import { AAMarkersSelector } from 'context/anticipatoryActionStateSlice';
+import { useAAMarkerScalePercent } from 'utils/map-utils';
+import LegendItemsList from 'components/MapView/Legends/LegendItemsList';
+import { Panel, leftPanelTabValueSelector } from 'context/leftPanelStateSlice';
+import {
+  dateRangeSelector,
+  mapSelector,
+} from '../../../context/mapStateSlice/selectors';
+import { useSafeTranslation } from '../../../i18n';
+import { MapDimensions } from './printImage.types';
+
+type Toggles = {
+  fullLayerDescription: boolean;
+  countryMask: boolean;
+  mapLabelsVisibility: boolean;
+  logoVisibility: boolean;
+  legendVisibility: boolean;
+};
+
+function ImagePreview({
+  classes,
+  open,
+  toggles,
+  mapDimensions,
+  footerHeight,
+  titleText,
+  titleRef,
+  footerTextSize,
+  footerText,
+  footerRef,
+  logoPosition,
+  titleHeight,
+  logoScale,
+  legendPosition,
+  legendScale,
+  invertedAdminBoundaryLimitPolygon,
+}: ImagePreviewProps) {
+  const { t } = useSafeTranslation();
+  const selectedMap = useSelector(mapSelector);
+  const dateRange = useSelector(dateRangeSelector);
+  const AAMarkers = useSelector(AAMarkersSelector);
+  const tabValue = useSelector(leftPanelTabValueSelector);
+  const printRef = useRef<HTMLDivElement>(null);
+  const northArrowRef = useRef<HTMLImageElement>(null);
+
+  const mapRef = React.useRef<MapRef>(null);
+
+  // Get the style and layers of the old map
+  const selectedMapStyle = selectedMap?.getStyle();
+
+  if (selectedMapStyle && !toggles.mapLabelsVisibility) {
+    // eslint-disable-next-line fp/no-mutation
+    selectedMapStyle.layers = selectedMapStyle?.layers.filter(
+      x => !x.id.includes('label'),
+    );
+  }
+
+  const updateScaleBarAndNorthArrow = React.useCallback(() => {
+    const elem = document.querySelector(
+      '.maplibregl-ctrl-scale',
+    ) as HTMLElement;
+
+    // this takes into account the watermark
+    const baseHeight = footerHeight || 20;
+
+    if (elem) {
+      // eslint-disable-next-line fp/no-mutating-assign
+      Object.assign(elem.style, {
+        position: 'absolute',
+        right: '10px',
+        bottom: `${baseHeight + 10}px`,
+        margin: 0,
+      });
+    }
+
+    if (northArrowRef.current) {
+      northArrowRef.current.style.bottom = `${baseHeight + 40}px`;
+    }
+  }, [footerHeight]);
+
+  React.useEffect(() => {
+    updateScaleBarAndNorthArrow();
+  }, [updateScaleBarAndNorthArrow]);
+
+  const { logo } = appConfig.header || {};
+  const scalePercent = useAAMarkerScalePercent(mapRef.current?.getMap());
+  const dateText = `${t('Publication date')}: ${getFormattedDate(
+    Date.now(),
+    'default',
+  )}${
+    dateRange.startDate
+      ? `. ${t('Layer selection date')}: ${getFormattedDate(
+          dateRange.startDate,
+          'default',
+        )}`
+      : ''
+  }.`;
+
+  return (
+    <div className={classes.previewContainer}>
+      <div
+        style={{
+          width: '100%',
+          height: '100%',
+        }}
+      >
+        <div
+          style={{
+            position: 'relative',
+            zIndex: 3,
+            border: '1px solid black',
+            height: `${mapDimensions.height}%`,
+            width: `${mapDimensions.width}%`,
+          }}
+        >
+          <div ref={printRef} className={classes.printContainer}>
+            <img
+              ref={northArrowRef}
+              style={{
+                position: 'absolute',
+                zIndex: 3,
+                width: '50px',
+                bottom: `${footerHeight + 40}px`,
+                right: '10px',
+              }}
+              src="./images/icon_north_arrow.png"
+              alt="northArrow"
+            />
+            {titleText && (
+              <div ref={titleRef}>
+                <Box
+                  display="inline-block"
+                  bgcolor="white"
+                  px={2}
+                  py={1}
+                  boxShadow={3}
+                  borderRadius={2}
+                  zIndex={2}
+                  position="absolute"
+                  top={0}
+                  left="50%"
+                  style={{ transform: 'translateX(-50%)' }}
+                >
+                  <Typography variant="h6">{titleText}</Typography>
+                </Box>
+              </div>
+            )}
+            {footerTextSize > 0 && (footerText || dateText) && (
+              <div
+                ref={footerRef}
+                className={classes.footerOverlay}
+                style={{
+                  fontSize: `${footerTextSize}px`,
+                }}
+              >
+                {footerText && (
+                  <Typography style={{ whiteSpace: 'pre-line' }}>
+                    {footerText}
+                  </Typography>
+                )}
+                {dateText && <Typography>{dateText}</Typography>}
+              </div>
+            )}
+            {logoPosition !== -1 && (
+              <img
+                style={{
+                  position: 'absolute',
+                  zIndex: 2,
+                  top: titleHeight + 8,
+                  height: 32 * logoScale,
+                  left: logoPosition % 2 === 0 ? '8px' : 'auto',
+                  right: logoPosition % 2 === 0 ? 'auto' : '8px',
+                  display: 'flex',
+                  justifyContent:
+                    logoPosition % 2 === 0 ? 'flex-start' : 'flex-end',
+                }}
+                src={logo}
+                alt="logo"
+              />
+            )}
+            {legendPosition !== -1 && (
+              <div
+                style={{
+                  position: 'absolute',
+                  zIndex: 2,
+                  top:
+                    titleHeight +
+                    (logoPosition === legendPosition ? 32 * logoScale : 0),
+                  left: legendPosition % 2 === 0 ? '8px' : 'auto',
+                  right: legendPosition % 2 === 0 ? 'auto' : '8px',
+                  display: 'flex',
+                  justifyContent:
+                    legendPosition % 2 === 0 ? 'flex-start' : 'flex-end',
+                  width: '20px',
+                  // Use transform scale to adjust size based on legendScale
+                  transform: `scale(${1 - legendScale})`,
+                }}
+              >
+                <LegendItemsList
+                  forPrinting
+                  listStyle={classes.legendListStyle}
+                  showDescription={toggles.fullLayerDescription}
+                />
+              </div>
+            )}
+            <div className={classes.mapContainer}>
+              {selectedMap && open && (
+                <MapGL
+                  ref={mapRef}
+                  dragRotate={false}
+                  // preserveDrawingBuffer is required for the map to be exported as an image
+                  preserveDrawingBuffer
+                  initialViewState={{
+                    longitude: selectedMap.getCenter().lng,
+                    latitude: selectedMap.getCenter().lat,
+                    zoom: selectedMap.getZoom(),
+                  }}
+                  onLoad={e => {
+                    e.target.addControl(
+                      new maplibregl.ScaleControl({}),
+                      'bottom-right',
+                    );
+                    updateScaleBarAndNorthArrow();
+                  }}
+                  mapStyle={selectedMapStyle || mapStyle.toString()}
+                  maxBounds={selectedMap.getMaxBounds() ?? undefined}
+                >
+                  {tabValue === Panel.AnticipatoryAction &&
+                    AAMarkers.map(marker => (
+                      <Marker
+                        key={`marker-${marker.district}`}
+                        longitude={marker.longitude}
+                        latitude={marker.latitude}
+                        anchor="center"
+                      >
+                        <div style={{ transform: `scale(${scalePercent})` }}>
+                          {marker.icon}
+                        </div>
+                      </Marker>
+                    ))}
+                  {toggles.countryMask && (
+                    <Source
+                      id="mask-overlay"
+                      type="geojson"
+                      data={invertedAdminBoundaryLimitPolygon}
+                    >
+                      <Layer
+                        id="mask-layer-overlay"
+                        type="fill"
+                        source="mask-overlay"
+                        layout={{}}
+                        paint={{
+                          'fill-color': '#000',
+                          'fill-opacity': 0.7,
+                        }}
+                      />
+                    </Source>
+                  )}
+                </MapGL>
+              )}
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+const styles = () =>
+  createStyles({
+    backdrop: {
+      position: 'absolute',
+    },
+    backdropWrapper: {
+      position: 'absolute',
+      top: 0,
+      left: 0,
+      zIndex: 2,
+      width: '100%',
+      height: '100%',
+      display: 'flex',
+      justifyContent: 'center',
+    },
+    printContainer: {
+      width: '100%',
+      height: '100%',
+    },
+    mapContainer: {
+      position: 'absolute',
+      top: 0,
+      left: 0,
+      height: '100%',
+      width: '100%',
+      zIndex: 1,
+    },
+    titleOverlay: {
+      position: 'absolute',
+      top: 0,
+      left: 0,
+      zIndex: 2,
+      color: 'black',
+      backgroundColor: 'white',
+      width: '100%',
+      textAlign: 'center',
+      fontSize: '1.25rem',
+      fontWeight: 600,
+      padding: '8px 0 8px 0',
+      borderBottom: `1px solid ${lightGrey}`,
+    },
+    footerOverlay: {
+      padding: '8px',
+      position: 'absolute',
+      bottom: 0,
+      left: 0,
+      zIndex: 3,
+      color: 'black',
+      backgroundColor: 'white',
+      width: '100%',
+      borderTop: `1px solid ${lightGrey}`,
+    },
+    legendListStyle: {
+      position: 'absolute',
+      top: '8px',
+      zIndex: 2,
+    },
+    sameRowToggles: {
+      display: 'flex',
+      flexDirection: 'row',
+      justifyContent: 'space-between',
+    },
+    previewContainer: {
+      display: 'flex',
+      flexDirection: 'column',
+      flexGrow: 1,
+    },
+  });
+
+export interface ImagePreviewProps extends WithStyles<typeof styles> {
+  open: boolean;
+  toggles: Toggles;
+  mapDimensions: MapDimensions;
+  footerHeight: number;
+  selectedBoundaries: AdminCodeString[];
+  setDownloadMenuAnchorEl: (anchorEl: HTMLElement | null) => void;
+  titleText: string;
+  titleRef: React.RefObject<HTMLDivElement>;
+  footerTextSize: number;
+  footerText: string;
+  footerRef: React.RefObject<HTMLDivElement>;
+  logoPosition: number;
+  titleHeight: number;
+  logoScale: number;
+  legendPosition: number;
+  legendScale: number;
+  invertedAdminBoundaryLimitPolygon: any;
+}
+
+export default withStyles(styles)(ImagePreview);
