+++ frontend/src/components/NavBar/PrintImage/printImage.types.ts
@@ -1,0 +1,9 @@
+export type Toggles = {
+  fullLayerDescription: boolean;
+  countryMask: boolean;
+  mapLabelsVisibility: boolean;
+  logoVisibility: boolean;
+  legendVisibility: boolean;
+};
+
+export type MapDimensions = { width: number; height: number };
